added more tests for kubernetes-client cache
