+++ backend/.nycrc
@@ -1,10 +1,10 @@
 {
   "all": true,
   "check-coverage": true,
-  "lines": 87,
-  "statements": 87,
-  "functions": 91,
-  "branches": 65,
+  "lines": 89,
+  "statements": 89,
+  "functions": 92,
+  "branches": 67,
   "include": [
     "lib/**/*.js"
   ],
